10-04-18

+++ src/main/java/com/example/repositories/EventRepository.java
@@ -6,11 +6,11 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.example.domain.Event;
 
-public interface EventRepository extends CrudRepository<Event, Long>{
+public interface EventRepository extends CrudRepository<Event, Integer>{
 
     List<Event> findAll();
 
-    Event findOne(Long id);
+    Event findOne(int id);
 
     Event save(Event event);
 }
